requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
